[change] 初回起動時の画面遷移変更
