+++ app/src/main/AndroidManifest.xml
@@ -21,6 +21,7 @@
         <activity android:name=".CreateRoomActivity"></activity>
         <activity android:name=".VideoActivity"></activity>
         <activity android:name=".SearchVideoActivity"></activity>
+        <activity android:name=".FirstStartActivity"></activity>
         <activity android:name=".SignInActivity"></activity>
         <activity android:name=".SignUpActivity"></activity>
         <meta-data android:name="developer_key" android:value="${developer_key}" />
