+++ app/src/main/java/com/example/atsushi/youtubesync/SignUpActivity.java
@@ -6,6 +6,7 @@ import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
 import android.view.View;
 import android.widget.EditText;
+
 import com.example.atsushi.youtubesync.json_data.User;
 import com.example.atsushi.youtubesync.server.SignUp;
 import com.example.atsushi.youtubesync.server.SignUpInterface;
@@ -16,8 +17,6 @@ import com.example.atsushi.youtubesync.server.SignUpInterface;
 
 public class SignUpActivity extends AppCompatActivity implements SignUpInterface {
 
-    private final int SIGN_IN_REQUEST_CODE = 100;
-    private SignUpActivity self = this;
     private static String TAG = SignUpActivity.class.getSimpleName();
 
     @Override
@@ -47,42 +46,13 @@ public class SignUpActivity extends AppCompatActivity implements SignUpInterface
                         }
                     }
                 });
-
-        findViewById(R.id.sign_up_link)
-                .setOnClickListener(new View.OnClickListener() {
-                    @Override
-                    public void onClick(View v) {
-                        Intent varIntent =
-                                new Intent(SignUpActivity.this, SignInActivity.class);
-                        startActivityForResult(varIntent, SIGN_IN_REQUEST_CODE);
-                    }
-                });
-    }
-
-    @Override
-    protected void onActivityResult(int requestCode, int resultCode, Intent intent) {
-        super.onActivityResult(requestCode, resultCode, intent);
-        if (resultCode == RESULT_OK && intent != null) {
-            if (requestCode == SIGN_IN_REQUEST_CODE) {
-                String res = intent.getStringExtra("access_token");
-                if (res != null) {
-                    new Token(this).setToken(res);
-                    Intent main = new Intent(SignUpActivity.this, MainActivity.class);
-                    startActivity(main);
-                }
-            }
-        }
     }
 
     @Override
     public void onSignedUp(final User user) {
-        runOnUiThread(new Runnable() {
-            public void run() {
-                new Token(self).setToken(user.access_token);
-                Intent main = new Intent(SignUpActivity.this, MainActivity.class);
-                startActivity(main);
-                finish();
-            }
-        });
+        Intent intent = new Intent();
+        intent.putExtra("access_token", user.access_token);
+        setResult(RESULT_OK, intent);
+        finish();
     }
 }
