+++ app/src/main/res/values/strings.xml
@@ -14,6 +14,5 @@
     <string name="password">Password</string>
     <string name="confirm_password">Password（確認）</string>
     <string name="new_register">新規登録</string>
-    <string name="already_get_account">既にアカウントを持っている方はこちら"</string>
     <string name="login_button">ログイン</string>
 </resources>
